pyproject.toml
[tool.poetry]
name = "agents-of-inference"
version = "0.1.0"
description = ""
authors = ["briancaffey"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.1"
langgraph = "^0.0.60"
langchain-openai = "^0.1.8"
jupyter = "^1.0.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pillow = "^10.3.0"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
